修复rm删除不存在的目录弹出提示的bug

+++ clean.ps1
@@ -2,4 +2,6 @@ $project_path=Split-Path -Parent $MyInvocation.MyCommand.Definition
 $project_name=Split-Path -Leaf $project_path
 $target_path=Join-Path $project_path "target"
 
-Remove-Item $target_path -Force -Recurse
+if(Test-Path -Path $target_path){
+    Remove-Item $target_path -Force -Recurse
+}
