+++ git_init.ps1
@@ -3,7 +3,9 @@ $project_name=Split-Path -Leaf $project_path
 $git_path=Join-Path -Path $project_path -ChildPath ".git"
 $user_name="fgwsz"
 
-Remove-Item $git_path -Force -Recurse
+if(Test-Path -Path $git_path){
+    Remove-Item $git_path -Force -Recurse
+}
 cd $project_path
 git init
 echo "# $repo_name" > README.md
